requirements.txt
google-adk>=0.4.0
google-genai
python-dotenv
pydantic
rich
beautifulsoup4
fake-useragent
